refactor: 패키지 분리 및 Dto 네이밍 변경
