+++ backend/src/main/java/com/digginroom/digginroom/admin/controller/dto/UploadRequest.java
@@ -1,10 +1,10 @@
-package com.digginroom.digginroom.controller;
+package com.digginroom.digginroom.admin.controller.dto;
 
 import com.digginroom.digginroom.domain.Genre;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record UploadRequestDto(
+public record UploadRequest(
         @NotBlank String title,
         @NotBlank String artist,
         @NotNull Genre superGenre,
