+++ backend/src/main/java/com/digginroom/digginroom/admin/service/UploadService.java
@@ -1,6 +1,6 @@
-package com.digginroom.digginroom.service;
+package com.digginroom.digginroom.admin.service;
 
-import com.digginroom.digginroom.controller.UploadRequestDto;
+import com.digginroom.digginroom.admin.controller.dto.UploadRequest;
 import com.digginroom.digginroom.domain.MediaSource;
 import com.digginroom.digginroom.domain.Room;
 import com.digginroom.digginroom.domain.Track;
@@ -17,7 +17,7 @@ public class UploadService {
     private final RoomRepository roomRepository;
 
     @Transactional
-    public void save(final UploadRequestDto request) {
+    public void save(final UploadRequest request) {
         MediaSource mediaSource = new MediaSource(request.youtubeVideoId());
         Track track = Track.builder()
                 .title(request.title())
